+++ app/services/bot_context_v2/commands/parsers/make_check.rb
@@ -12,6 +12,8 @@ module BotContextV2
             act.on('--dis [TEXT]', Integer) { |text=0| result[:adv] = text.to_i * -1 }
             act.on('--bonus [TEXT]', Integer) { |text=0| result[:bonus] = text.to_i }
             act.on('--penalty [TEXT]', Integer) { |text=0| result[:bonus] = text.to_i * -1 }
+            act.on('--target [TEXT]', Integer) { |text=0| result[:target] = text.to_i }
+            act.on('--expertise [TEXT]', Integer) { |text=0| result[:expertise] = text.to_i }
           end
           parser.parse! arguments
           result